Add refresh to RBAC list Fix unit tests
